1.4.1-added simple system report

+++ database/migrations/2022_11_13_171849_create_posts_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->string('title');
             $table->string('slug');
             $table->string('thumbnail_src')->nullable();
-            $table->unsignedBigInteger('song_id')->nullable();
-            $table->unsignedBigInteger('artist_id')->nullable();
+            $table->unsignedInteger('song_id')->nullable();
+            $table->unsignedInteger('artist_id')->nullable();
             $table->enum('type', ['OP', 'ED'])->nullable();
             $table->enum('status', ['stagged', 'published'])->nullable();
             $table->string('suffix')->nullable();
